add skibidi level to game setup
