+++ assets/js/platformer3x/GameSetup.js
@@ -3,50 +3,18 @@ import GameEnv from './GameEnv.js';
 import GameLevel from './GameLevel.js';
 // To build GameLevels, each contains GameObjects from below imports
 import Background from './Background.js'
-import BackgroundParallax from './BackgroundParallax.js';
-import BackgroundTransitions from './BackgroundTransitions.js';
-import BackgroundSnow from './BackgroundSnow.js';
-import Platform from './Platform.js';
-import JumpPlatform from './JumpPlatform.js';
-import PlayerHills from './PlayerHills.js';
-import PlayerWinter from './PlayerWinter.js';
-import PlayerMini from './PlayerMini.js';
-import PlayerMiniHogwarts from './PlayerMiniHogwarts.js';
-import PlayerQuidditch from './PlayerQuidditch.js';
-import Goomba from './EnemyGoomba.js';
-import FlyingGoomba from './FlyingGoomba.js';
-import BlockPlatform from './BlockPlatform.js';
-import SpawnPlatform from './PlatformSpawn.js';
-import MovingPlatform from './PlatformMoving.js'
-import Mushroom from './Mushroom.js';
-import MagicBeam from './MagicBeam.js';
-import ChocoFrog from './ChocoFrog.js';
-import Coin from './Coin.js';
 import GameControl from './GameControl.js';
-import Owl from './FlyingOwl.js';
-import Snowman from './EnemySnowman.js';
-import Cerberus from './EnemyCerberus.js';
-import PlayerGreece from './PlayerGreece.js';
-import FinishLine from './FinishLine.js';
-import Lava from './Lava.js';
-import Dragon from './FlyingDragon.js';
-import Star from './Star.js';
-import Dementor from './FlyingDementor.js';
-import Draco from './EnemyDraco.js';
-import Boss from './Boss.js';
-import skibidiTitan from './SkibidiTitan.js';
-import Laser from './Laser.js';
-import SkibidiToilet from './SkibidiToilet.js';
-import PlayerSkibidi from './PlayerSkibidi.js';
-import Tree from './Tree.js';
-import Jellyfish from './FlyingJellyfish.js';
-import Penguin from './EnemyPenguin.js';
-import PlayerIce from './PlayerIce.js';
-import FlyingIsland from './PlatformFlyingIsland.js';
-import PlayerBaseOneD from './PlayerBaseOneD.js';
-import PlayerZombie from './PlayerZombie.js';
-import BossItem from './BossItem.js';
-import PlayerBoss from './PlayerBoss.js';
+import GameSet from './GameSet.js';
+import GameSetterHills from './GameSetterHills.js';
+import GameSetterWater from './GameSetterWater.js';
+import GameSetterGreece from './GameSetterGreece.js';
+import GameSetterGreeceMini from './GameSetterGreeceMini.js';
+import GameSetterQuidditch from './GameSetterQuidditch.js';
+import GameSetterHogwarts from './GameSetterHogwarts.js';
+import GameSetterWinter from './GameSetterWinter.js';
+import GameSetterWinterIce from './GameSetterWinterIce.js';
+import GameSetterBoss from './GameSetterBoss.js';
+import GameSetterSkibidi from './GameSetterSkibidi.js';
 
 //test comment
 
@@ -220,484 +188,12 @@ const GameSetup = {
   */
 
   assets: {
-    obstacles: {
-      tube: { 
-        src: "/images/platformer/obstacles/tube.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5},
-        width: 300,
-        height: 300,
-        scaleSize: 100,
-      },
-      tubeU: { 
-        src: "/images/platformer/obstacles/blue-tube-up.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5},
-        width: 300,
-        height: 300,
-        scaleSize: 100,
-      },
-      tubeD: { 
-        src: "/images/platformer/obstacles/blue-tube.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5},
-        width: 300,
-        height: 300,
-        scaleSize: 100,
-      },
-      toilet: { 
-        src: "/images/platformer/obstacles/toilet.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5}
-      },
-      laser: {
-        src: "/images/platformer/obstacles/laser.png",
-        hitbox: { widthPercentage: 0.1, heightPercentage: 0.5}
-      },
-      cabin: {
-        src: "/images/platformer/obstacles/cabin.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 },
-        width: 300,
-        height: 300,
-        scaleSize: 150,
-      },
-      iceberg: {
-        src: "/images/platformer/obstacles/iceberg.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 },
-        width: 300,
-        height: 300,
-        scaleSize: 150,
-      },
-      chest: {
-        src: "/images/platformer/obstacles/Chest.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 },
-        width: 300,
-        height: 300,
-        scaleSize: 100,
-      },
-      flag: {
-        src: "/images/platformer/obstacles/flag.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 },
-        width: 300,
-        height: 300,
-        scaleSize: 120,
-      },
-      vbucks: { src: "/images/platformer/obstacles/vbucks.png"},
-      coin: { src: "/images/platformer/obstacles/coin.png" },
-      snowflake: { src: "/images/platformer/obstacles/snowflake.png" },
-      star: { src: "/images/platformer/obstacles/star.png" },
-      snitch: { src: "/images/platformer/obstacles/snitch.png" },
-      whompingwillow: {
-        src: "/images/platformer/obstacles/whompingwillowtree.png",
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 }
-      },
-    },
-    platforms: {
-      grass: { src: "/images/platformer/platforms/grass.png" },
-      narwhalfloor: { src: "/images/platformer/platforms/narwhalfloor.png" },
-      sand: { src: "/images/platformer/platforms/sand.png" },
-      snowyfloor: { src: "/images/platformer/platforms/snowyfloor.png" },
-      snowywood: { src: "/images/platformer/platforms/snowywood.png" },
-      alien: { src: "/images/platformer/platforms/alien.png" },
-      bricks: { src: "/images/platformer/platforms/brick_wall.png" },
-      lava: { src: "/images/platformer/platforms/lava.jpg" },
-      sandstone: { src: "/images/platformer/platforms/sandstone.png" },
-      cobblestone: { src: "/images/platformer/platforms/cobblestone.png" },
-      yellowpattern: { src: "/images/platformer/platforms/yellowtowerpattern.jpg" },
-      yellowredpattern: { src: "/images/platformer/platforms/yellowredpattern.jpg" },
-      lionpattern: { src: "/images/platformer/platforms/lionpattern.jpg" },
-      stone: { src: "/images/platformer/platforms/stone.jpg"}, 
-      turf: { src: "/images/platformer/platforms/turf.png" },
-      island: { src: "/images/platformer/platforms/island.png" },
-      island: { src: "/images/platformer/platforms/island.png" },
-      block: { src: "/images/platformer/platforms/brick_block.png" }, //MAY need 3 new variables: sizeRatio, widthRatio, and heightRatio
-      itemBlock: {
-        src: "/images/platformer/platforms/mario_block_spritesheet_v2.png",
-        sizeRatio: 83.2,
-        widthRatio: 0.5,
-        heightRatio: 1.0,
-        width: 204,
-        height: 204,
-        scaleSize: 80,
-        speedRatio: 0.7,
-        hitbox: { widthPercentage: 0.4, heightPercentage: -0.2 }
-      }
-    },
     backgrounds: {
-      boss: { src: "/images/platformer/backgrounds/BossBackground.png", parallaxSpeed: 0.4, moveOnKeyAction: true },
       start: { src: "/images/platformer/backgrounds/home.png" },
-      hills: { src: "/images/platformer/backgrounds/hills.png", parallaxSpeed: 0.4, moveOnKeyAction: true },
-      greece: { src: "/images/platformer/backgrounds/greek.png" },
-      mountains: { src: "/images/platformer/backgrounds/mountains.jpg", parallaxSpeed: 0.1, moveOnKeyAction: true },
-      clouds: { src: "/images/platformer/backgrounds/clouds.png", parallaxSpeed: 0.5 },
-      water: { src: "/images/platformer/backgrounds/water.png" },
-      fish: { src: "/images/platformer/backgrounds/school-fish.png", parallaxSpeed: -0.5 },
-      complete3: { src: "/images/platformer/backgrounds/skibidiCompletion.png" },
-      desert: {src: "/images/platformer/backgrounds/desertbg.png"},
-      reef: { src: "/images/platformer/backgrounds/reef.png" },
-      quidditch: { src: "/images/platformer/backgrounds/quidditch2.jpg" },
-      miniHogwarts: { src: "/images/platformer/backgrounds/miniHogwarts.png"}, 
-      space: { src: "/images/platformer/backgrounds/planet.jpg" },
-      castles: { src: "/images/platformer/backgrounds/castles.png" },
-      winter: { src: "/images/platformer/backgrounds/winter.png", parallaxSpeed: 0.4, moveOnKeyAction: true },
-      snow: { src: "/images/platformer/backgrounds/snowfall.png" },
-      icewater: { src: "/images/platformer/backgrounds/icewater.png", parallaxSpeed: 0.4, moveOnKeyAction: true},
-      narwhal: { src: "/images/platformer/backgrounds/narwhal.png", parallaxSpeed: 2 },
-      mini: { src: "/images/platformer/backgrounds/mini.png" },
-      devil: {src: "/images/platformer/backgrounds/devil.png", parallaxSpeed: 2 },
     },
     transitions: {
-      loading: { src: "/images/platformer/transitions/greenscreen.png" },
-      hillsEnd: { src: "/images/platformer/transitions/hillsEnd.png" },
-      winterEnd: { src: "/images/platformer/transitions/winterEnd.png" },
-      greeceEnd: { src: "/images/platformer/transitions/greeceEnd.png" },
-      waterEnd: { src: "/images/platformer/transitions/waterEnd.png" },
-      quidditchEnd: { src: "/images/platformer/transitions/quidditchEnd.png" },
       miniEnd: { src: "/images/platformer/transitions/miniEnd.png" },
-      iceminiEnd: { src: "/images/platformer/transitions/IceMinigameEnd.png"},
     },
-    players: {
-      mario: {
-        src: "/images/platformer/sprites/mario.png",
-        width: 256,
-        height: 256,
-        scaleSize: 80,
-        speedRatio: 0.7,
-        idle: {
-          left: { row: 1, frames: 15 },
-          right: { row: 0, frames: 15 },
-        },
-        walk: {
-          left: { row: 3, frames: 7 },
-          right: { row: 2, frames: 7 },
-        },
-        run: {
-          left: { row: 5, frames: 15 },
-          right: { row: 4, frames: 15 },
-        },
-        jump: {
-          left: { row: 11, frames: 15 },
-          right: { row: 10, frames: 15 },
-        },
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-      escaper: {
-        src: "/images/platformer/sprites/escaper.png",
-        width: 130,
-        height: 140,
-        scaleSize: 150,
-        speedRatio: 0.7,
-        animationSpeed: 8,
-        ///animationspeed:6
-        idle: {
-            left: { row: 0, frames: 5 },
-            right: { row: 0, frames: 5},
-        },
-        walk: {
-            left: { row: 1, frames: 6 },
-            right: { row: 1, frames: 6 },
-        },
-        run: {
-            left: { row: 2, frames: 7 },
-            right: { row: 2, frames: 7 },
-        },
-        jump: {
-            left: { row: 3, frames: 8 },
-            right: { row: 3, frames: 8 },
-        },
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-      whitemario: {
-        src: "/images/platformer/sprites/white_mario.png",
-        width: 256,
-        height: 256,
-        scaleSize: 80,
-        speedRatio: 0.7,
-        idle: {
-          left: { row: 1, frames: 15 },
-          right: { row: 0, frames: 15 },
-        },
-        walk: {
-          left: { row: 3, frames: 7 },
-          right: { row: 2, frames: 7 },
-        },
-        run: {
-          left: { row: 5, frames: 15 },
-          right: { row: 4, frames: 15 },
-        },
-        jump: {
-          left: { row: 11, frames: 15 },
-          right: { row: 10, frames: 15 },
-        },
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-      monkey: {
-        src: "/images/platformer/sprites/monkey.png",
-        width: 40,
-        height: 40,
-        scaleSize: 80,
-        speedRatio: 0.7,
-        wa: { row: 9, min: 8, frames: 15 },
-        wd: { row: 9, min: 0, frames: 7 },
-        a: { row: 1, frames: 15, idleFrame: { column: 7, frames: 0 } },
-        s: { row: 12, frames: 15 },
-        d: { row: 0, frames: 15, idleFrame: { column: 7, frames: 0 } }
-      },
-      knight: {
-        src: "/images/platformer/sprites/knight.png",
-        width: 128,
-        height: 128,
-        scaleSize: 120,
-        speedRatio: 0.7,
-        idle: {
-          left: { row: 1, frames: 23 },
-          right: { row: 0, frames: 23 },
-        },
-        walk: {
-          left: { row: 7, frames: 20 },
-          right: { row: 6, frames: 20 },
-        },
-        run: {
-          left: { row: 5, frames: 23 },
-          right: { row: 4, frames: 23 },
-        },
-        jump: {
-          left: { row: 3, frames: 23 },
-          right: { row: 2, frames: 23 },
-        },
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      }, harry: {
-        src: "/images/platformer/sprites/harryanimation3.png",
-        width: 32,
-        height: 32,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        idle: {
-          left: { row: 1, frames: 0 },
-          right: { row: 2, frames: 0 },
-        },
-        walk: {
-          left: { row: 1, frames: 5 },
-          right: { row: 2, frames: 5 },
-        },
-        run: {
-          left: { row: 1, frames: 5 },
-          right: { row: 2, frames: 5 },
-        },
-        jump: {
-          left: { row: 1, frames: 0 },
-          right: { row: 2, frames: 0 },
-        },
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-      lopez: {
-        src: "/images/platformer/sprites/lopezanimation.png",
-        width: 46,
-        height: 52.5,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        wa: { row: 1, frames: 3 }, // Up-Left Movement 
-        wd: { row: 2, frames: 3 }, // Up-Right Movement
-        idle: { row: 6, frames: 1, idleFrame: { column: 1, frames: 0 } },
-        a: { row: 1, frames: 3, idleFrame: { column: 1, frames: 0 } }, // Left Movement
-        s: { row: 1, frames: 3 }, // Stop the movement 
-        d: { row: 2, frames: 3, idleFrame: { column: 1, frames: 0 } }, // Right Movement 
-        runningLeft: { row: 5, frames: 3, idleFrame: { column: 1, frames: 0 } },
-        runningRight: { row: 4, frames: 3, idleFrame: { column: 1, frames: 0 } },
-      },
-      zombie: { //one direction player
-        src: "/images/platformer/sprites/zombie.png",
-        width: 130,
-        height: 70,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        idle: { row: 2, frames: 11, idleFrame: { column: 1, frames: 0 } },
-        walk: { row: 3, frames: 11 }, // default - right Movement
-        run: { row: 3, frames: 11 }, // default - right Movement
-        jump: { row: 3, frames: 11 }, // default - right Movement
-        attack: { row: 4, min: 6,frames: 11 }, 
-        jumpAttack : { row: 6, frames: 11 }, 
-        death : { row: 11, frames: 11 }, 
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-      zombie: { //one direction player
-        src: "/images/platformer/sprites/zombie.png",
-        width: 130,
-        height: 70,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        idle: { row: 2, frames: 11, idleFrame: { column: 1, frames: 0 } },
-        walk: { row: 3, frames: 11 }, // default - right Movement
-        run: { row: 3, frames: 11 }, // default - right Movement
-        jump: { row: 3, frames: 11 }, // default - right Movement
-        attack: { row: 4, min: 6,frames: 11 }, 
-        jumpAttack : { row: 6, frames: 11 }, 
-        death : { row: 11, frames: 11 }, 
-        hitbox: { widthPercentage: 0.3, heightPercentage: 0.8 }
-      },
-    },
-    enemies: {
-      goomba: {
-        src: "/images/platformer/sprites/goomba.png",
-        width: 448,
-        height: 452,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        xPercentage: 0.6,
-        hitbox: { widthPercentage: 0.0, heightPercentage: 0.2 }
-      },
-      Snowman: {
-        src: "/images/platformer/sprites/snowman.png",
-        width: 308,
-        height: 327,
-        scaleSize: 60,
-        speedRatio: 0.7,
-        xPercentage: 0.6,
-        hitbox: { widthPercentage: 0.0, heightPercentage: 0.2 },
-        left: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Left Movement
-        idle: { row: 0, frames: 0 }, // Stop the movement 
-        right: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Right Movement 
-      },
-      Penguin: {
-        src: "/images/platformer/sprites/penguin.png",
-        width: 240,
-        height: 290,
-        scaleSize: 80,
-        speedRatio: 0.6,
-        xPercentage: 0.6,
-        hitbox: { widthPercentage: 0.0, heightPercentage: 0.2 },
-        left: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Left Movement
-        idle: { row: 0, frames: 0 }, // Stop the movement 
-        right: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Right Movement 
-      },
-      Owl: {
-        src: "/images/platformer/sprites/owl.png",
-        width: 499,
-        height: 500,
-        scaleSize: 60,
-        speedRatio: 0.8,
-      },
-      Jellyfish: {
-        src: "/images/platformer/sprites/jellyfish.png",
-        width: 499, 
-        height: 500,
-        scaleSize: 90,
-        speedRatio: 0.8,
-      },  
-      flyingGoomba: {
-        src: "/images/platformer/sprites/flying-goomba.png",
-        width: 448,
-        height: 452,
-        scaleSize: 60,
-        speedRatio: 0.7,
-      },
-      mushroom: {
-        src: "/images/platformer/platforms/mushroom.png",
-        width: 200,
-        height: 180,
-        hitbox: { widthPercentage: 0.0, heightPercentage: 0.2 }
-      },
-      magicBeam: {
-        src: "/images/platformer/platforms/magic_beam.png",
-        width: 450,
-        height: 400,
-        hitbox: { widthPercentage: 0.5, heightPercentage: 0.5 }
-      },
-      chocoFrog: {
-        src: "/images/platformer/platforms/Chocolatefrog.jpg",
-        width: 200,
-        height: 200,
-        hitbox: { widthPercentage: 0.0, heightPercentage: 0.0 }
-      },
-      alien: {
-        src: "/images/platformer/sprites/alien.png",
-        width: 444,
-        height: 640,
-        scaleSize: 60,
-        speedRatio: 0.85,
-      },
-      flyingUFO: {
-        src: "/images/platformer/sprites/flying-ufo.png",
-        width: 1920,
-        height: 1166,
-        scaleSize: 150,
-        speedRatio: 0.9,
-      },
-      cerberus: {
-        src: "/images/platformer/sprites/cerberus.png",
-        width: 103,
-        height: 103,
-        scaleSize: 80,
-        speedRatio: 0.85,
-        left: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Left Movement
-        idle: { row: 0, frames: 0 }, // Stop the movement 
-        right: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Right Movement 
-      },
-      skibidiToilet: {
-        src: "/images/platformer/sprites/skibidiEnemy.png",
-        width: 529,
-        height: 884,
-        scaleSize: 60,
-        speedRatio: 0.85,
-      },
-      skibidiTitan: {
-        src: "/images/platformer/sprites/skibidiTItan.png",
-        width: 529,
-        height: 884,
-        scaleSize: 1500,
-        speedRatio: 0.85,
-      },
-      dragon: {
-        src: "/images/platformer/sprites/dragon.png",
-        width: 152,
-        height: 119,
-        scaleSize: 60,
-        speedRatio: 0.7,
-      }, dementor: {
-        src: "/images/platformer/sprites/dementor2.png",
-        width: 400,
-        height: 400,
-        scaleSize: 80,
-        speedRatio: 0.7,
-      },
-      draco: {
-        src: "/images/platformer/sprites/dracomalfoy.png",
-        width: 301,
-        height: 261,
-        scaleSize: 80,
-        speedRatio: 0.7,
-        xPercentage: 0.6,
-        left: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Left Movement
-        idle: { row: 0, frames: 0 }, // Stop the movement 
-        right: { row: 0, frames: 0, idleFrame: { column: 0, frames: 0 } }, // Right Movement 
-      },
-      boss: {
-        src: "/images/platformer/sprites/boss.png",
-        width: 64,
-        height: 64,
-        scaleSize: 320,
-        speedRatio: 0.6,
-        animationSpeed: 6,
-        idleL: { row: 9, frames: 0, idleFrame: { column: 1, frames: 0 } },
-        idleR: { row: 11, frames: 0, idleFrame: { column: 1, frames: 0 } },
-        left: { row: 9, frames: 8, idleFrame: { column: 7, frames: 0 } },
-        right: { row: 11, frames: 8, idleFrame: { column: 7, frames: 0 } },
-        attackL: { row: 13, frames: 5 },
-        attackR: { row: 15, frames: 5 },
-        death: { row: 20, frames: 5 },
-      },
-      narwhalboss: {
-        src: "/images/platformer/sprites/narwhal_boss.png",
-        width: 64,
-        height: 64,
-        scaleSize: 320,
-        speedRatio: 0.6,
-        animationSpeed: 6,
-        idleL: { row: 0, frames: 4, idleFrame: { column: 1, frames: 0 } },
-        idleR: { row: 1, frames: 4, idleFrame: { column: 1, frames: 0 } },
-        left: { row: 2, frames: 4, idleFrame: { column: 1, frames: 0 } },
-        right: { row: 3, frames: 4, idleFrame: { column: 1, frames: 0 } },
-        attackL: { row: 4, frames: 4 },
-        attackR: { row: 5, frames: 4 },
-        death: { row: 6, frames: 4 },
-      },
-    }
   },
 
   /*  ==========================================
@@ -770,493 +266,45 @@ const GameSetup = {
     // Home Screen Background added to the GameEnv, "passive" means complementary, not an interactive level..
     new GameLevel({ tag: "home", callback: this.homeScreenCallback, objects: homeGameObjects, passive: true });
 
-    // Check local storage for the difficulty mode set
-    let difficulty = localStorage.getItem("difficulty") || "easy";
-
-    // If difficulty is not set (null or undefined), set it to a default value
-    // if (!difficulty) {
-    //     difficulty = "normal"; // Set default difficulty to "normal" or any other suitable value
-    // }
-
-    // Hills Game Level defintion...
-    const allHillsGameObjects = [
-      { name: 'mountains', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.mountains },
-      { name: 'clouds', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.clouds },
-      { name: 'hills', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.hills },
-      { name: 'grass', id: 'floor', class: Platform, data: this.assets.platforms.grass },
-      { name: 'blocks', id: 'jumpPlatform', class: MovingPlatform, data: this.assets.platforms.block, xPercentage: 0.2, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: MovingPlatform, data: this.assets.platforms.block, xPercentage: 0.2368, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: MovingPlatform, data: this.assets.platforms.block, xPercentage: 0.2736, yPercentage: 0.85 },
-      { name: 'blocks', id: 'wall', class: SpawnPlatform, data: this.assets.platforms.block, xPercentage: 0.6, yPercentage: 1 },
-      { name: 'itemBlock', id: 'jumpPlatform', class: JumpPlatform, data: this.assets.platforms.itemBlock, xPercentage: 0.4, yPercentage: 0.65 }, //item block is a platform
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.5, yPercentage: 1, minPosition: 0.05 },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.4, yPercentage: 1, minPosition: 0.05, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.3, yPercentage: 1, minPosition: 0.05, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.2, yPercentage: 1, minPosition: 0.05, difficulties: ["hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.1, yPercentage: 1, minPosition: 0.05, difficulties: ["impossible"] },
-      { name: 'goombaSpecial', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.75, yPercentage: 1, minPosition: 0.5 }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'goombaSpecial', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.95, yPercentage: 1, minPosition: 0.5, difficulties: ["hard", "impossible"] }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["hard", "impossible"] },
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["impossible"] },
-      { name: 'mushroom', id: 'mushroom', class: Mushroom, data: this.assets.enemies.mushroom, xPercentage: 0.49 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.1908, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.2242, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.2575, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.5898, yPercentage: 0.900 },
-      { name: 'mario', id: 'player', class: PlayerHills, data: this.assets.players.mario },
-      { name: 'tube', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tube, xPercentage: 0.85, yPercentage: 0.65 },
-      { name: 'loading', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.loading },
-    ];
-    let hillsGameObjects = allHillsGameObjects.filter(obj => !obj.difficulties || obj.difficulties.includes(difficulty));
     // Hills Game Level added to the GameEnv ...
-    new GameLevel({ tag: "hills", callback: this.playerOffScreenCallBack, objects: hillsGameObjects });
+    var hillsGameObjects = new GameSet(GameSetterHills.assets, GameSetterHills.objects, path);
+    new GameLevel({ tag: "hills", callback: this.playerOffScreenCallBack, objects: hillsGameObjects.getGameObjects() });
 
-
-    // Greece Game Level definition...
-    const greeceGameObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'greece', id: 'background', class: Background, data: this.assets.backgrounds.greece },
-      { name: 'grass', id: 'platform', class: Platform, data: this.assets.platforms.grass },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.2, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.2368, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.2736, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.3104, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.3472, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.384, yPercentage: 0.76 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.4208, yPercentage: 0.70 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.5090, yPercentage: 0.64 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.5642, yPercentage: 0.34 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.5274, yPercentage: 0.34 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.4906, yPercentage: 0.34 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 1 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.94 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.88 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.76 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.70 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.64 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.58 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.52 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.46 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.40 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.34 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.28 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.22 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6368, yPercentage: 0.64 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.16 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.1 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.6, yPercentage: 0.06 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 1 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.94 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.88 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.82 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.76 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.70 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.64 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.58 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.52 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.46 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.40 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.34 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.28 },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.22 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.16 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.1 },
-      //{ name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sandstone, xPercentage: 0.75, yPercentage: 0.06 },
-      { name: 'cerberus', id: 'cerberus', class: Cerberus, data: this.assets.enemies.cerberus, xPercentage: 0.2, minPosition: 0.09, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'cerberus', id: 'cerberus', class: Cerberus, data: this.assets.enemies.cerberus, xPercentage: 0.2, minPosition: 0.09, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'cerberus', id: 'cerberus', class: Cerberus, data: this.assets.enemies.cerberus, xPercentage: 0.5, minPosition: 0.3, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'cerberus', id: 'cerberus', class: Cerberus, data: this.assets.enemies.cerberus, xPercentage: 0.7, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },//this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'dragon', id: 'dragon', class: Dragon, data: this.assets.enemies.dragon, xPercentage: 0.5, minPosition: 0.05 },
-      { name: 'knight', id: 'player', class: PlayerGreece, data: this.assets.players.knight },
-      { name: 'flyingIsland', id: 'flyingIsland', class: FlyingIsland, data: this.assets.platforms.island, xPercentage: 0.82, yPercentage: 0.55 },
-      { name: 'tubeU', id: 'minifinishline', class: FinishLine, data: this.assets.obstacles.tubeU, xPercentage: 0.66, yPercentage: 0.71 },
-      { name: 'flag', id: 'finishline', class: FinishLine, data: this.assets.obstacles.flag, xPercentage: 0.875, yPercentage: 0.21 },
-      { name: 'flyingIsland', id: 'flyingIsland', class: FlyingIsland, data: this.assets.platforms.island, xPercentage: 0.82, yPercentage: 0.55 },
-      { name: 'hillsEnd', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.hillsEnd },
-      { name: 'lava', id: 'lava', class: Lava, data: this.assets.platforms.lava, xPercentage: 0, yPercentage: 1 },
-      { name: 'lava', id: 'lava', class: Lava, data: this.assets.platforms.lava, xPercentage: 0, yPercentage: 1 },
-    ];
     // Greece Game Level added to the GameEnv ...
-    new GameLevel({ tag: "ancient greece", callback: this.playerOffScreenCallBack, objects: greeceGameObjects });
+    var greeceGameObjects = new GameSet(GameSetterGreece.assets, GameSetterGreece.objects, path);
+    new GameLevel({ tag: "ancient greece", callback: this.playerOffScreenCallBack, objects: greeceGameObjects.getGameObjects() });
 
-
-    const miniGameObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'mini', id: 'background', class: Background, data: this.assets.backgrounds.mini },
-      // { name: 'rock', id: 'platform', class: Platform, data: this.assets.platforms.rock },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.59, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.6268, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3368, yPercentage: 0.35 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3368, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.4684, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.6, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.6368, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3736, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3736, yPercentage: 0.4334 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3736, yPercentage: 0.5167 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.3736, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.4104, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.4472, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.484, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.5208, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.5576, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.5576, yPercentage: 0.5167 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.5576, yPercentage: 0.4334 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.5576, yPercentage: 0.35 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8576, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8576, yPercentage: 0.5167 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8576, yPercentage: 0.4334 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8576, yPercentage: 0.35 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8576, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.8208, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.784, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.7472, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.7104, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.6736, yPercentage: 0.6 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.1736, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.1736, yPercentage: 0.4334 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.1736, yPercentage: 0.5167 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lava, xPercentage: 0.1736, yPercentage: 0.6 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.28, yPercentage: 0.25 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.32, yPercentage: 0.25 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.29, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.33, yPercentage: 0.75 },
-      { name: 'star', id: 'star', class: Star, data: this.assets.obstacles.star, xPercentage: 0.4584, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.40, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.42, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.44, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.46, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.48, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.5, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.59, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.63, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.58, yPercentage: 0.25 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.62, yPercentage: 0.25 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.6475, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.6675, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.6875, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.7075, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.7275, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.7475, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.7675, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.7875, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.8075, yPercentage: 0.5 },
-      { name: 'knight', id: 'player', class: PlayerMini, data: this.assets.players.knight },
-      { name: 'tubeD', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tubeD, xPercentage: 0, yPercentage: 0.052 },
-      { name: 'tubeU', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tubeU, xPercentage: 0.85, yPercentage: 0.646 },
-      { name: 'greeceEnd', id: 'background', class: BackgroundTransitions,  data: this.assets.transitions.greeceEnd },
-    ];
     // Space Game Level added to the GameEnv ...
-    new GameLevel({ tag: "mini", callback: this.playerOffScreenCallBack, objects: miniGameObjects });
-
+    var greeceMiniGameObjects = new GameSet(GameSetterGreeceMini.assets, GameSetterGreeceMini.objects, path);
+    new GameLevel({ tag: "mini greece", callback: this.playerOffScreenCallBack, objects: greeceMiniGameObjects.getGameObjects() });
 
     // Under Water Game Level defintion...
-    const allWaterGameObjects = [
-      { name: 'water', id: 'background', class: Background, data: this.assets.backgrounds.water },
-      { name: 'fish', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.fish },
-      { name: 'reef', id: 'background', class: Background, data: this.assets.backgrounds.reef },
-      { name: 'sand', id: 'floor', class: Platform, data: this.assets.platforms.sand },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.block, xPercentage: 0.2, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.block, xPercentage: 0.2368, yPercentage: 0.85 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.block, xPercentage: 0.2736, yPercentage: 0.85 },
-      { name: 'blocks', id: 'wall', class: BlockPlatform, data: this.assets.platforms.block, xPercentage: 0.6, yPercentage: 1 },
-      { name: 'itemBlock', id: 'jumpPlatform', class: JumpPlatform, data: this.assets.platforms.itemBlock, xPercentage: 0.4, yPercentage: 0.65 }, //item block is a platform
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.5, yPercentage: 1, minPosition: 0.05 },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.4, yPercentage: 1, minPosition: 0.05, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.3, yPercentage: 1, minPosition: 0.05, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.2, yPercentage: 1, minPosition: 0.05, difficulties: ["hard", "impossible"] },
-      { name: 'goomba', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.1, yPercentage: 1, minPosition: 0.05, difficulties: ["impossible"] },
-      { name: 'goombaSpecial', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.75, yPercentage: 1, minPosition: 0.5 }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'goombaSpecial', id: 'goomba', class: Goomba, data: this.assets.enemies.goomba, xPercentage: 0.95, yPercentage: 1, minPosition: 0.5, difficulties: ["hard", "impossible"] }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["hard", "impossible"] },
-      { name: 'flyingGoomba', id: 'flyingGoomba', class: FlyingGoomba, data: this.assets.enemies.flyingGoomba, xPercentage: 0.9, minPosition: 0.5, difficulties: ["impossible"] },
-      { name: 'mushroom', id: 'mushroom', class: Mushroom, data: this.assets.enemies.mushroom, xPercentage: 0.49 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.1908, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.2242, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.2575, yPercentage: 0.75 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.coin, xPercentage: 0.5898, yPercentage: 0.900 },
-      { name: 'mario', id: 'player', class: PlayerHills, data: this.assets.players.mario },
-      { name: 'Chest', id: 'finishline', class: FinishLine, data: this.assets.obstacles.chest, xPercentage: 0.85, yPercentage: 0.65 },
-      { name: 'miniEnd', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.miniEnd },
-    ];
-    let waterGameObjects = allWaterGameObjects.filter(obj => !obj.difficulties || obj.difficulties.includes(difficulty));
-    // Water Game Level added to the GameEnv ...
-    new GameLevel({ tag: "water", callback: this.playerOffScreenCallBack, objects: waterGameObjects });
-
-    // Quidditch Game Level definition...
-    const quidditchGameObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'quidditch', id: 'background', class: Background, data: this.assets.backgrounds.quidditch },
-      { name: 'turf', id: 'platform', class: Platform, data: this.assets.platforms.turf },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.1, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.14, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.18, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.22, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.22, yPercentage: 0.69 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.71 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.61 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.33 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.23 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.30, yPercentage: 0.13 },
-      { name: 'blocks', id: 'jumpPlatform', class: SpawnPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.34, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: SpawnPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: SpawnPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.42, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.57 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.37 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.27 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.38, yPercentage: 0.17 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: MovingPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.626, yPercentage: 0.92 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.663, yPercentage: 0.586 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.cobblestone, xPercentage: 0.700, yPercentage: 0.586 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowpattern, xPercentage: 0.456, yPercentage: 1.08 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowredpattern, xPercentage: 0.456, yPercentage: 0.985 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowpattern, xPercentage: 0.456, yPercentage: 0.89 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lionpattern, xPercentage: 0.456, yPercentage: 0.795 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowpattern, xPercentage: 0.456, yPercentage: 0.7 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowredpattern, xPercentage: 0.456, yPercentage: 0.605 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.yellowpattern, xPercentage: 0.456, yPercentage: 0.51 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.lionpattern, xPercentage: 0.456, yPercentage: 0.415 },
-
-      { name: 'draco', id: 'draco', class: Draco, data: this.assets.enemies.draco, xPercentage: 0.3, minPosition: 0.05, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'draco', id: 'draco', class: Draco, data: this.assets.enemies.draco, xPercentage: 0.5, minPosition: 0.3, difficulties: ["normal", "hard", "impossible"] },
-      /**{ name: 'draco', id: 'draco', class: Draco, data: this.assets.enemies.draco, xPercentage: 0.75, minPosition: 0.5, difficulties: ["normal", "hard", "impossible"] }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event */
-      { name: 'dementor', id: 'dementor', class: Dementor, data: this.assets.enemies.dementor, xPercentage: 0.5, minPosition: 0.05 },
-      { name: 'dementor', id: 'dementor', class: Dementor, data: this.assets.enemies.dementor, xPercentage: 0.9, minPosition: 0.5 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.095, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.135, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.175, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.375, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.409, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.295, yPercentage: 0.46 },
-
-      { name: 'chocoFrog', id: 'chocoFrog', class: ChocoFrog, data: this.assets.enemies.chocoFrog, xPercentage: 0.25, yPercentage: 0.3 },
-
-      { name: 'magicBeam', id: 'magicBeam', class: MagicBeam, data: this.assets.enemies.magicBeam, xPercentage: 0.623, yPercentage: 0.72 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.656, yPercentage: 0.46 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.611, yPercentage: 0.46 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.700, yPercentage: 0.46 },
-
-      { name: 'harry', id: 'player', class: PlayerQuidditch, data: this.assets.players.harry },
-      { name: 'tube', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tube, xPercentage: 0.85, yPercentage: 0.7 },
-      { name: 'tubeU', id: 'minifinishline', class: FinishLine, data: this.assets.obstacles.tubeU, xPercentage: 0.69, yPercentage: 0.71 },
-      { name: 'waterEnd', id: 'background', class: BackgroundTransitions,  data: this.assets.transitions.waterEnd },
-    ];
+    // Hills Game Level added to the GameEnv ...
+    var waterGameObjects = new GameSet(GameSetterWater.assets, GameSetterWater.objects, path);
+    new GameLevel({ tag: "water", callback: this.playerOffScreenCallBack, objects: waterGameObjects.getGameObjects() });
 
     // Quidditch Game Level added to the GameEnv ...
-    new GameLevel({ tag: "quidditch", callback: this.playerOffScreenCallBack, objects: quidditchGameObjects });
-
-    const miniHogwartsObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'miniHogwarts', id: 'background', class: Background, data: this.assets.backgrounds.miniHogwarts },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.009, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.058, yPercentage: 0.66 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.1, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.14, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.18, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.22, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.26, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.30, yPercentage: 0.47 },
-      
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.14, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.18, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.22, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.26, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.30, yPercentage: 0.81 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: MovingPlatform, data: this.assets.platforms.stone, xPercentage: 0.44, yPercentage: 0.92 },
-      { name: 'magicBeam', id: 'magicBeam', class: MagicBeam, data: this.assets.enemies.magicBeam, xPercentage: 0.37, yPercentage: 0.61 },
-
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.48, yPercentage: 0.64 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.48, yPercentage: 0.54 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.48, yPercentage: 0.44 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.48, yPercentage: 0.34 },
-      
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.6, yPercentage: 0.66 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.56, yPercentage: 0.5 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.64, yPercentage: 0.81 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.68, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.72, yPercentage: 0.47 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.stone, xPercentage: 0.76, yPercentage: 0.47 },
-      
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.55, yPercentage: 0.38 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.636, yPercentage: 0.699 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.672, yPercentage: 0.368 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.71, yPercentage: 0.368 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.75, yPercentage: 0.368 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.056, yPercentage: 0.56 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.15, yPercentage: 0.24 },
-      
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.14, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.18, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.22, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.26, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.43, yPercentage: 0.82 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.47, yPercentage: 0.24 },
-
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.snitch, xPercentage: 0.85, yPercentage: 0.81 },
-      
-
-      { name: 'harry', id: 'player', class: PlayerMiniHogwarts, data: this.assets.players.harry },
-      { name: 'tubeD', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tubeD, xPercentage: 0, yPercentage: 0.052 },
-      { name: 'tubeU', id: 'minifinishline', class: FinishLine, data: this.assets.obstacles.tubeU, xPercentage: 0.85, yPercentage: 0.7 },
-    ];
+    var quidditchGameObjects = new GameSet(GameSetterQuidditch.assets, GameSetterQuidditch.objects, path);
+    new GameLevel({ tag: "quidditch", callback: this.playerOffScreenCallBack, objects: quidditchGameObjects.getGameObjects() });
 
     // miniHogwarts Game Level added to the GameEnv ...
-    new GameLevel({ tag: "mini hogwarts", callback: this.playerOffScreenCallBack, objects: miniHogwartsObjects });
-
-    const winterObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'winter', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.winter },
-      { name: 'snow', id: 'background', class: BackgroundSnow, data: this.assets.backgrounds.snow },
-      { name: 'snowyfloor', id: 'platform', class: Platform, data: this.assets.platforms.snowyfloor },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2368, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2736, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3104, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3472, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.384, yPercentage: 0.74 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.4208, yPercentage: 0.66 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5090, yPercentage: 0.56 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5090, yPercentage: 0.48 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5090, yPercentage: 0.40 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5090, yPercentage: 0.32 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.69, yPercentage: 0.76 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.655, yPercentage: 0.68 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.62, yPercentage: 0.68 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.72, yPercentage: 0.76 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.755, yPercentage: 1 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.755, yPercentage: 0.92 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.755, yPercentage: 0.84 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.625, yPercentage: 0.92 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.625, yPercentage: 1 },
-      { name: 'snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.2100, yPercentage: 0.72 },
-      { name: 'snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.2619, yPercentage: 0.72 },
-      { name: 'snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.3136, yPercentage: 0.72 },
-      { name: 'owl', id: 'owl', class: Owl, data: this.assets.enemies.Owl, xPercentage: 0.3, minPosition: 0.05 },
-      { name: 'owl', id: 'owl', class: Owl, data: this.assets.enemies.Owl, xPercentage: 0.8, minPosition: 0.05 },
-      { name: 'snowman', id: 'snowman', class: Snowman, data: this.assets.enemies.Snowman, xPercentage: 0.2, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'snowman', id: 'snowman', class: Snowman, data: this.assets.enemies.Snowman, xPercentage: 0.35, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'snowman', id: 'snowman', class: Snowman, data: this.assets.enemies.Snowman, xPercentage: 0.5, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'mario', id: 'player', class: PlayerWinter, data: this.assets.players.whitemario },
-      { name: 'cabin', id: 'finishline', class: FinishLine, data: this.assets.obstacles.cabin, xPercentage: 0.85, yPercentage: 0.603 },
-      { name: 'tubeU', id: 'minifinishline', class: FinishLine, data: this.assets.obstacles.tubeU, xPercentage: 0.69, yPercentage: 0.71 },
-      { name: 'quidditchEnd', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.quidditchEnd },
-    ];
+    var miniHogwartsGameObjects = new GameSet(GameSetterHogwarts.assets, GameSetterHogwarts.objects, path);
+    new GameLevel({ tag: "mini hogwarts", callback: this.playerOffScreenCallBack, objects: miniHogwartsGameObjects.getGameObjects() });
 
     // Winter MiniGame Level added to the GameEnv ...
-    new GameLevel({ tag: "winter", callback: this.playerOffScreenCallBack, objects: winterObjects });
+    var winterGameObjects = new GameSet(GameSetterWinter.assets, GameSetterWinter.objects, path);
+    new GameLevel({ tag: "winter", callback: this.playerOffScreenCallBack, objects: winterGameObjects.getGameObjects() });
     
-    const iceminiObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'icewater', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.icewater },
-      { name: 'narwhal', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.narwhal },
-      { name: 'narwhalfloor', id: 'platform', class: Platform, data: this.assets.platforms.narwhalfloor },
-      { name: 'sandstone', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2368, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.2736, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3104, yPercentage: 0.82 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3472, yPercentage: 0.74 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.59, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.6268, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.332, yPercentage: 0.35 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.3736, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.4104, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.4472, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.484, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5208, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.snowywood, xPercentage: 0.5576, yPercentage: 0.6 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.325, yPercentage: 0.25 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.29, yPercentage: 0.25 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.40, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.42, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.44, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.46, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.48, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.5, yPercentage: 0.5 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.58, yPercentage: 0.25 },
-      { name: 'Snowflake', id: 'coin', class: Coin, data: this.assets.obstacles.snowflake, xPercentage: 0.62, yPercentage: 0.25 },
-      { name: 'jellyfish', id: 'jellyfish', class: Jellyfish, data: this.assets.enemies.Jellyfish, xPercentage: 0.2, minPosition: 0.05 },
-      { name: 'jellyfish', id: 'jellyfish', class: Jellyfish, data: this.assets.enemies.Jellyfish, xPercentage: 0.8, minPosition: 0.05 },
-      { name: 'penguin', id: 'penguin', class: Penguin, data: this.assets.enemies.Penguin, xPercentage: 0.2, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'penguin', id: 'penguin', class: Penguin, data: this.assets.enemies.Penguin, xPercentage: 0.35, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'penguin', id: 'penguin', class: Penguin, data: this.assets.enemies.Penguin, xPercentage: 0.5, minPosition: 0.1, difficulties: ["normal", "hard", "impossible"] },
-      { name: 'mario', id: 'player', class: PlayerIce, data: this.assets.players.whitemario },
-      { name: 'tubeD', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tubeD, xPercentage: 0, yPercentage: 0.052 },
-      { name: 'iceberg', id: 'finishline', class: FinishLine, data: this.assets.obstacles.iceberg, xPercentage: 0.85, yPercentage: 0.603 },
-      { name: 'winterEnd', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.winterEnd },
-    ];
     // IceMiniGame Game Level added to the GameEnv ...
-    new GameLevel({ tag: "icemini", callback: this.playerOffScreenCallBack, objects: iceminiObjects });
+    var winterIceGameObjects = new GameSet(GameSetterWinterIce.assets, GameSetterWinterIce.objects, path);
+    new GameLevel({ tag: "icemini", callback: this.playerOffScreenCallBack, objects: winterIceGameObjects.getGameObjects() });
 
-    //Skibidi Toilet Level
-    const skibidiGameObjects = [
-      // GameObject(s), the order is important to z-index...
-      { name: 'desert', id: 'background', class: Background, data: this.assets.backgrounds.desert },
-      //{ name: 'clouds', id: 'background', class: BackgroundClouds, data: this.assets.backgrounds.clouds },
-      { name: 'skibidiTitan', id: 'skibidiTitan', class: skibidiTitan, data: this.assets.enemies.skibidiTitan, xPercentage:  0.35, yPercentage: 0.5, minPosition: 0.5 }, 
-      { name: 'sand', id: 'platform', class: Platform, data: this.assets.platforms.sand },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.2, yPercentage: 1 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.4, yPercentage: 0.6 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.325, yPercentage: 0.8 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.2, yPercentage: 0.5 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.225, yPercentage: 0.5 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.0, yPercentage: 0.5 } ,
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.025, yPercentage: 0.5 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.025, yPercentage: 0.5 },
-      { name: 'blocks', id: 'jumpPlatform', class: BlockPlatform, data: this.assets.platforms.sand, xPercentage: 0.5, yPercentage: 0.5 },
-      ///{ name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.vbucks, xPercentage: 0.475, yPercentage: 0.5 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.vbucks, xPercentage: 0.325, yPercentage: 0.7 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.vbucks, xPercentage: -0.0125, yPercentage: 0.4 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.vbucks, xPercentage: 0.0125, yPercentage: 0.4 },
-      { name: 'coin', id: 'coin', class: Coin, data: this.assets.obstacles.vbucks, xPercentage: 0.0325, yPercentage: 0.4 },
-      { name: 'SkibidiToilet', id: 'SkibidiToilet', class: SkibidiToilet, data: this.assets.enemies.skibidiToilet, xPercentage:  0.3, minPosition: 0.07 },
-      { name: 'SkibidiToilet', id: 'SkibidiToilet', class: SkibidiToilet, data: this.assets.enemies.skibidiToilet, xPercentage:  0.5, minPosition: 0.3 },
-      { name: 'SkibidiToilet', id: 'SkibidiToilet', class: SkibidiToilet, data: this.assets.enemies.skibidiToilet, xPercentage:  0.75, minPosition: 0.5 }, //this special name is used for random event 2 to make sure that only one of the Goombas ends the random event
-      { name: 'escaper', id: 'player', class: PlayerSkibidi, data: this.assets.players.escaper  },
-      { name: 'laser', id: 'Laser', class: Laser, data: this.assets.obstacles.laser, xPercentage:  0.75, yPercentage: 0.5 },
-      { name: 'toiletTube', id: 'toiletEnd', class: Tree, data: this.assets.obstacles.toilet },
-      { name: 'complete3', id: 'background', class: BackgroundTransitions,  data: this.assets.backgrounds.complete3 },
-    ];
+    // Boss Game Level added to the GameEnv ...
+    var bossGameObjects = new GameSet(GameSetterBoss.assets, GameSetterBoss.objects, path);
+    new GameLevel({ tag: "boss", callback: this.playerOffScreenCallBack, objects: bossGameObjects.getGameObjects() });
 
-    new GameLevel( {tag: "skibidi", callback: this.playerOffScreenCallBack, objects: skibidiGameObjects} );
-
-
-    const bossGameObjects = [
-      { name: 'bossbackground', id: 'background', class: BackgroundParallax, data: this.assets.backgrounds.boss },
-      { name: 'devil', id: 'devil', class:BackgroundParallax, data: this.assets.backgrounds.devil},
-      { name: 'boss', id: 'boss', class: Boss, data: this.assets.enemies.boss, xPercentage: 0.5, minPosition: 0.3 },
-      { name: 'boss1', id: 'boss', class: Boss, data: this.assets.enemies.boss, xPercentage: 0.3, minPosition: 0.07 },
-      { name: 'mario', id: 'player', class: PlayerBoss, data: this.assets.players.mario },
-      { name: 'zombie', id: 'player', class: PlayerZombie, data: this.assets.players.zombie },
-      { name: 'tube', id: 'finishline', class: FinishLine, data: this.assets.obstacles.tube, xPercentage: 0.85, yPercentage: 0.65 },
-      { name: 'itemBlock', id: 'jumpPlatform', class: BossItem, data: this.assets.platforms.itemBlock, xPercentage: 0.2, yPercentage: 0.65 }, //item block is a platform
-      { name: 'mario', id: 'player', class: PlayerBoss, data: this.assets.players.mario },
-      { name: 'zombie', id: 'player', class: PlayerZombie, data: this.assets.players.zombie },
-      { name: 'grass', id: 'platform', class: Platform, data: this.assets.platforms.grass },
-      { name: 'iceminiEnd', id: 'background', class: BackgroundTransitions, data: this.assets.transitions.iceminiEnd },
-    ];
-
-    new GameLevel({ tag: "boss", callback: this.playerOffScreenCallBack, objects: bossGameObjects });
+    const skibidiGameObjects = new GameSet(GameSetterSkibidi.assets, GameSetterSkibidi.objects, path)
+    new GameLevel({ tag: "skibdi", callback: this.playerOffScreenCallBack, objects: skibidiGameObjects.getGameObjects() })
 
     // Game Over Level definition...
     const endGameObjects = [
